KAFKA-6185; Remove channels from explictlyMutedChannels set when closed
This memory leak could eventually lead to an OutOfMemoryError. This
was particularly likely in case of down conversions as the leaked
channels would hold on to the record batch (which is only loaded
into the heap in case of down conversions).

Author: Rajini Sivaram <rajinisivaram@googlemail.com>

Reviewers: Ismael Juma <ismael@juma.me.uk>

Closes #4193 from rajinisivaram/KAFKA-6185-oom

(cherry picked from commit 0653a895f51492a8687aaba0dc7a43b21fca2a56)
Signed-off-by: Ismael Juma <ismael@juma.me.uk>