[DO NOT MERGE] 4.1.1 Hotfix for SSL connection-related issues (#198)
* KAFKA-3702: Change log level of SSL close_notify failure (#5397)

SslTransportLayer currently closes the SSL engine and logs a warning if close_notify message canot be sent because the remote end closed its connection. This tends to fill up broker logs, especially when using clients which close connections immediately. Since this log entry is not very useful anyway, it would be better to log at debug level.

Reviewers: Jason Gustafson <jason@confluent.io>, Ismael Juma <ismael@juma.me.uk>

* KAFKA-7454: Use lazy allocation for SslTransportLayer buffers and null them on close (#5713)

Lazy allocation helps when there are a large number of connections
that have been accepted, but where no data has been received from
the clients. Each buffer is often around 16k (max TLS record size).

Nulling the buffers should not make a difference in the current
implementation since we release the reference to the channel
and transport layer after we close them, but it's a good practice
to release medium/large buffers after `close` is called.

Reviewers: Jun Rao <junrao@gmail.com>, Ismael Juma <ismael@juma.me.uk>

* KAFKA-7453: Expire registered channels not selected within idle timeout (#5712)

Reviewers: Jun Rao <junrao@gmail.com>. Ismael Juma <ismael@juma.me.uk>

* KAFKA-7304: Limit connection accept rate to prioritize other channels

* Log close of idle connections at `info`

* Checkstyle

* Fix processor retry logic to advance to the next processor on retry

* Add test for connection rate limit

* Wrap socketChannel close in try block