MINOR: Fix log message when tasks directory is cleaned manually (#9262)
Currently when a task directory is cleaned manually the message
for the state dir cleaner is logged instead of the message for
the manual cleanup. This is because the code checks the elapsed
time since the last update before it checks whether the cleanup
is a manual call. This commit changes the order of the checks.

Reviewers: A. Sophie Blee-Goldman <sophie@confluent.io>, Chia-Ping Tsai <chia7712@gmail.com>, Matthias J. Sax <mjsax@apache.org>, Walker Carlson <wcarlson@confluent.io>, John Roesler <vvcephei@apache.org>