Add LocalQuorumsTestEnv, fix LocalLogManager
Add a MetaLogManager object to QuorumController.

Create the LocalQuorumsTestEnv class to test QuorumController in
conjunection with LocalLogManager.

Rework the MetaLogManager API to require an initialization function.
Otherwise we run into circular dependencies when trying to call the
controller's constructor.  This also means reworking LocalLogManager so
that it only starts its threads inside LocalLogManager#initialize(),
rather than in the constructor.

Fix a case where a LocalLogManager follower could throw an exception due
to reading from the log before the leader had finished writing to it.
In other words, fix a case where we weren't handling EOF correctly.