Get Kip500ClusterTest working
* change port types from int16 to int32 to avoid seeing negative port
numbers when ports that don't fit in a signed int16 are used.

* BrokerLifecycleManager should fill in the features field when sending
a registration request.

* Kip500Broker: avoid creating an instance of kafka.cluster.Broker.
This was only ever used to convert to an AuthorizerServerInfo.  So just
create one of those directly.

* Pass the advertised listeners in BrokerLifecycleManager#start rather
than computing them from the config.  Computing from the config
doesn't work when we are doing a JUnit test and the port is set up
dynamically (via random assignment).